refactor: split CLI into internal modules
