+++ quillan/cli_app/handlers/routing.py
@@ -1,0 +1,175 @@
+"""Scan-routing command handler."""
+
+from __future__ import annotations
+
+import argparse
+from datetime import datetime, timezone
+from pathlib import Path
+
+from pds_core.pds1 import Pds1PayloadError, parse_pds1_payload
+from pds_core.workspace import WorkspaceRootError, resolve_workspace_root
+
+from quillan.cli_app.output import (
+    print_evidence_filing_review,
+    print_route_failure_review,
+    print_routed_evidence,
+)
+from quillan.evidence_filing import (
+    EvidenceFilingError,
+    file_routed_response_evidence,
+)
+from quillan.route_planning import (
+    DecodedResponsePage,
+    RouteFailure,
+    plan_decoded_response_page_route,
+)
+from quillan.routing_review import (
+    RoutingReviewError,
+    preserve_evidence_filing_error_for_review,
+    preserve_route_failure_for_review,
+    preserve_routing_failure_for_review,
+)
+
+
+def handle_route_scan(args: argparse.Namespace) -> int:
+    """Route one source scan from caller-supplied decoded PDS1 text."""
+    source_file: Path = args.source_file
+    payload_text: str = args.payload
+    intake_timestamp = datetime.now(timezone.utc)
+    try:
+        workspace_root = resolve_workspace_root()
+    except WorkspaceRootError as error:
+        print(f"Error: could not resolve the PDS workspace: {error}")
+        return 1
+    except Exception as error:
+        print(f"Error: unexpected workspace resolution failure: {error}")
+        return 1
+
+    if not _validate_source_file(source_file):
+        return 1
+
+    try:
+        payload = parse_pds1_payload(payload_text)
+    except Pds1PayloadError as error:
+        return _preserve_payload_parse_failure(
+            workspace_root,
+            source_file=source_file,
+            payload_text=payload_text,
+            error=error,
+            created_at=intake_timestamp,
+        )
+    except Exception as error:
+        print(f"Error: unexpected PDS1 payload parsing failure: {error}")
+        return 1
+
+    decoded_page = DecodedResponsePage(
+        module=payload.module,
+        document_type=payload.metadata.get("doc"),
+        class_id=payload.class_id,
+        assignment_id=payload.assignment_id,
+        student_id=payload.student_id,
+        page_number=payload.page,
+        raw_payload=payload_text,
+    )
+
+    try:
+        route_result = plan_decoded_response_page_route(
+            workspace_root,
+            decoded_page,
+        )
+        if isinstance(route_result, RouteFailure):
+            review_record = preserve_route_failure_for_review(
+                workspace_root,
+                route_failure=route_result,
+                source_filename=source_file.name,
+                created_at=intake_timestamp,
+            )
+            print_route_failure_review(route_result, review_record)
+            return 0
+
+        try:
+            filed_evidence = file_routed_response_evidence(
+                workspace_root,
+                route_plan=route_result,
+                source_file_path=source_file,
+                intake_timestamp=intake_timestamp,
+            )
+        except EvidenceFilingError as error:
+            review_record = preserve_evidence_filing_error_for_review(
+                workspace_root,
+                error=error,
+                route_plan=route_result,
+                source_filename=source_file.name,
+                created_at=intake_timestamp,
+            )
+            print_evidence_filing_review(error, review_record)
+            return 0
+    except RoutingReviewError as error:
+        print(f"Error: could not preserve scan routing failure for review: {error}")
+        return 1
+    except Exception as error:
+        print(f"Error: unexpected route-scan failure: {error}")
+        return 1
+
+    print_routed_evidence(filed_evidence)
+    return 0
+
+
+def _validate_source_file(source_file: Path) -> bool:
+    """Return whether the selected scan is an existing readable file."""
+    try:
+        if not source_file.is_file():
+            print(
+                "Error: source file must be an existing regular file: "
+                f"{source_file}"
+            )
+            return False
+        with source_file.open("rb"):
+            pass
+    except OSError as error:
+        print(f"Error: source file is not readable: {error}")
+        return False
+    return True
+
+
+def _preserve_payload_parse_failure(
+    workspace_root: Path,
+    *,
+    source_file: Path,
+    payload_text: str,
+    error: Pds1PayloadError,
+    created_at: datetime,
+) -> int:
+    """Preserve malformed payload context without inventing route identity."""
+    try:
+        review_record = preserve_routing_failure_for_review(
+            workspace_root,
+            failure_category="payload_invalid",
+            failure_message=str(error),
+            source_filename=source_file.name,
+            module="quillan",
+            detected_payload=payload_text,
+            module_details={
+                "failure_origin": "payload_parse",
+                "parse_error": str(error),
+            },
+            created_at=created_at,
+        )
+    except RoutingReviewError as review_error:
+        print(
+            "Error: payload was invalid and could not be preserved for review: "
+            f"{review_error}"
+        )
+        return 1
+    except Exception as unexpected_error:
+        print(
+            "Error: unexpected failure while preserving invalid payload: "
+            f"{unexpected_error}"
+        )
+        return 1
+
+    print("Quillan response page was not routed; preserved for review.")
+    print(f"Reason: {error}")
+    print("Category: payload_invalid")
+    print(f"Review record: {review_record.failure_metadata_relative_path}")
+    return 0
